+++ src/extension.ts
@@ -2,12 +2,14 @@ import * as vscode from 'vscode';
 import { browseRow, freezeCmd, freezeManual, installCmd, installManual, installRow } from './commands';
 
 export function activate(context: vscode.ExtensionContext) {
-	context.subscriptions.push(installCmd);
-	context.subscriptions.push(freezeCmd);
-	context.subscriptions.push(installManual);
-	context.subscriptions.push(freezeManual);
-	context.subscriptions.push(installRow);
-	context.subscriptions.push(browseRow);
+	context.subscriptions.push(
+		installCmd,
+		freezeCmd,
+		installManual,
+		freezeManual,
+		installRow,
+		browseRow,
+	);
 }
 
 export function deactivate() {}