+++ src/web/extension.ts
@@ -2,7 +2,9 @@ import * as vscode from 'vscode';
 import { browseRow } from './commands';
 
 export function activate(context: vscode.ExtensionContext) {
-    context.subscriptions.push(browseRow);
+    context.subscriptions.push(
+        browseRow,
+    );
 }
 
 export function deactivate() {}